agcalls/my_projects/agcalls /Dockerfiles_wo_models/requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
gradio==5.31.0
librosa==0.11.0
soundfile==0.13.1
noisereduce==3.0.3
pydub==0.25.1
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.3.0
numpy==1.24.4
pandas==2.0.3
transformers==4.52.3
torch==2.7.0
onnxruntime==1.16.0
sentence-transformers==4.1.0
supervisor==4.2.5
celery==5.4.0
redis==5.0.8
python-jose==3.3.0
pytest==7.4.0